+++ js/main.js
@@ -1,0 +1,2 @@
+import './until.js';
+import { allPhotos } from './data.js';
